fuck you mac

+++ Assets/Demos/FunctionDev/CharacterControllerDemo/PlayerAccelerateCapability.cs
@@ -67,7 +67,7 @@ namespace CharacterControllerDemo
 
     public class PlayerAccelerateOnGroundCapability : PlayerAccelerateCapability
     {
-        protected override uint GetTickOrderInGroup() => PlayerMovementTickOrder.AccelerateOnGround;
+        protected override uint GetTickOrderInGroup() => (uint)PlayerMovementTickOrder.AccelerateOnGround;
 
         protected override float GetAcceleration() => mPlayerMovementComponent.AccelerationOnGround;
 
@@ -80,7 +80,7 @@ namespace CharacterControllerDemo
 
     public class PlayerAccelerateOnAirCapability : PlayerAccelerateCapability
     {
-        protected override uint GetTickOrderInGroup() => PlayerMovementTickOrder.AccelerateOnAir;
+        protected override uint GetTickOrderInGroup() => (uint)PlayerMovementTickOrder.AccelerateOnAir;
 
         protected override float GetAcceleration() => mPlayerMovementComponent.AccelerationOnAir;
 
@@ -94,7 +94,7 @@ namespace CharacterControllerDemo
     public class PlayerAccelerateWhileOverspeedOnGroundCapability : PlayerAccelerateCapability
     {
         protected override uint GetTickOrderInGroup() =>
-            PlayerMovementTickOrder.AccelerateWhileOverspeedOnGround;
+			(uint)PlayerMovementTickOrder.AccelerateWhileOverspeedOnGround;
 
         protected override float GetAcceleration() =>
             mPlayerMovementComponent.AccelerationWhileOverspeedOnGround;
@@ -109,7 +109,7 @@ namespace CharacterControllerDemo
     public class PlayerAccelerateWhileOverspeedOnAirCapability : PlayerAccelerateCapability
     {
         protected override uint GetTickOrderInGroup() =>
-            PlayerMovementTickOrder.AccelerateWhileOverspeedOnAir;
+			(uint)PlayerMovementTickOrder.AccelerateWhileOverspeedOnAir;
 
         protected override float GetAcceleration() =>
             mPlayerMovementComponent.AccelerationWhileOverspeedOnAir;
